# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GitHub Release 缓存和更新检查服务",
        "title": "Update Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8001",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "返回服务名称和版本",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "获取服务信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.RootResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/check-update": {
            "get": {
                "description": "根据客户端版本号判断是否需要更新",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "检查客户端是否有新版本",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"v1.0.0\"",
                        "description": "客户端当前版本号",
                        "name": "version",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateCheckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/download/{version}/{filename}": {
            "get": {
                "description": "从缓存或 GitHub 下载指定版本的文件",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "download"
                ],
                "summary": "下载指定版本的文件",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"v1.0.0\"",
                        "description": "版本号",
                        "name": "version",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"app-linux-amd64.tar.gz\"",
                        "description": "文件名",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/redirect/domains": {
            "get": {
                "description": "从 GitHub 获取 domains.json 并返回",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redirect"
                ],
                "summary": "获取域名列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/resources": {
            "get": {
                "description": "返回按平台分类的构建文件列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "获取构建资源列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ResourcesResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/version": {
            "get": {
                "description": "返回最新版本的完整信息，包括版本号、发布说明、资源列表等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "获取最新版本详情",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/version.Info"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/webhook": {
            "post": {
                "description": "接收 GitHub release 事件，自动更新版本信息和缓存",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhook"
                ],
                "summary": "GitHub Webhook 回调",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"release\"",
                        "description": "GitHub 事件类型",
                        "name": "X-GitHub-Event",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "GitHub 签名",
                        "name": "X-Hub-Signature-256",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.WebhookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.BuildInfo": {
            "type": "object",
            "properties": {
                "architecture": {
                    "type": "string"
                },
                "download_link": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_type": {
                    "type": "string"
                },
                "file_url": {
                    "type": "string"
                },
                "upload_time": {
                    "type": "string"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "错误信息"
                }
            }
        },
        "handler.ResourcesResponse": {
            "type": "object",
            "properties": {
                "builds": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/handler.BuildInfo"
                        }
                    }
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "handler.RootResponse": {
            "type": "object",
            "properties": {
                "app": {
                    "type": "string",
                    "example": "orange-service"
                },
                "version": {
                    "type": "string",
                    "example": "1.0.0"
                }
            }
        },
        "handler.UpdateCheckResponse": {
            "type": "object",
            "properties": {
                "download_url": {
                    "type": "string",
                    "example": "https://example.com"
                },
                "latest_version": {
                    "type": "string",
                    "example": "v1.2.0"
                },
                "release_notes": {
                    "type": "string",
                    "example": "Bug fixes and improvements"
                },
                "update_available": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "handler.WebhookResponse": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string",
                    "example": "not a release event"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                },
                "version": {
                    "type": "string",
                    "example": "v1.0.0"
                }
            }
        },
        "version.Asset": {
            "type": "object",
            "properties": {
                "download_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "version.Info": {
            "type": "object",
            "properties": {
                "assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/version.Asset"
                    }
                },
                "published_at": {
                    "type": "string"
                },
                "release_notes": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    }
}